Visual changes

+++ api_lib/src/main/java/oriolvillaret/com/api_connection_lib/models/Run.java
@@ -27,6 +27,6 @@ public class Run {
     public class Times {
         //in seconds
         @SerializedName("realtime_t")
-        public Integer realtime_t;
+        public Double realtime_t;
     }
 }
